indexer: refactor indexer handlers
