+++ backend/indexer/internal/infrastructure/database/proof_sets.go
@@ -4,59 +4,51 @@ import (
 	"context"
 	"fmt"
 	"math/big"
-	"pdp-explorer-indexer/internal/processor"
-	"time"
+
+	"pdp-explorer-indexer/internal/models"
 )
 
 // StoreProofSet stores a proof set record with version control
-func (p *PostgresDB) StoreProofSet(ctx context.Context, proofSet *processor.ProofSet) error {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	// Get the ID of the previous version to maintain history
-	var previousID *int64
-	err = tx.QueryRow(ctx, `
-		SELECT id FROM proof_sets 
-		WHERE set_id = $1 
-		ORDER BY block_number DESC 
-		LIMIT 1
-	`, proofSet.SetId).Scan(&previousID)
-	if err != nil {
-		// If no previous version exists, that's fine
-		previousID = nil
-	}
-
-	// Insert the new version
-	_, err = tx.Exec(ctx, `
+func (p *PostgresDB) StoreProofSet(ctx context.Context, proofSet *models.ProofSet) error {
+	_, err := p.pool.Exec(ctx, `
 		INSERT INTO proof_sets (
 			set_id, owner, listener_addr, total_faulted_periods, total_data_size,
 			total_roots, total_fee_paid, last_proven_epoch, next_challenge_epoch,
-			total_transactions, is_active, block_number, block_hash,
-			previous_id
+			total_transactions, is_active, block_number, block_hash, created_at, updated_at, total_proved_roots
 		) VALUES (
-			$1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14
+			$1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16
 		)
+		ON CONFLICT (set_id, block_number) DO UPDATE SET 
+			owner = EXCLUDED.owner,
+			listener_addr = EXCLUDED.listener_addr,
+			total_faulted_periods = EXCLUDED.total_faulted_periods,
+			total_data_size = EXCLUDED.total_data_size,
+			total_roots = EXCLUDED.total_roots,
+			total_fee_paid = EXCLUDED.total_fee_paid,
+			last_proven_epoch = EXCLUDED.last_proven_epoch,
+			next_challenge_epoch = EXCLUDED.next_challenge_epoch,
+			total_transactions = EXCLUDED.total_transactions,
+			is_active = EXCLUDED.is_active,
+			block_hash = EXCLUDED.block_hash,
+			updated_at = EXCLUDED.updated_at,
+			total_proved_roots = EXCLUDED.total_proved_roots
 	`, proofSet.SetId, proofSet.Owner, proofSet.ListenerAddr, proofSet.TotalFaultedPeriods,
 		proofSet.TotalDataSize, proofSet.TotalRoots, proofSet.TotalFeePaid,
 		proofSet.LastProvenEpoch, proofSet.NextChallengeEpoch, proofSet.TotalTransactions,
-		proofSet.IsActive, proofSet.BlockNumber, proofSet.BlockHash, previousID)
+		proofSet.IsActive, proofSet.BlockNumber, proofSet.BlockHash, proofSet.CreatedAt, proofSet.UpdatedAt, proofSet.TotalProvedRoots)
 	if err != nil {
 		return fmt.Errorf("failed to insert proof set: %w", err)
 	}
 
-	return tx.Commit(ctx)
+	return nil
 }
 
 // FindProofSet finds a proof set by its set_id, optionally including historical versions
-func (p *PostgresDB) FindProofSet(ctx context.Context, setId int64, includeHistory bool) ([]*processor.ProofSet, error) {
+func (p *PostgresDB) FindProofSet(ctx context.Context, setId int64, includeHistory bool) ([]*models.ProofSet, error) {
 	query := `
-		SELECT id, set_id, owner, listener_addr, total_faulted_periods, total_data_size,
+		SELECT id, set_id, owner, listener_addr, total_faulted_periods, total_proved_roots, total_data_size,
 			   total_roots, total_fee_paid, last_proven_epoch, next_challenge_epoch,
-			   total_transactions, is_active, block_number, block_hash,
-			   previous_id
+			   total_transactions, is_active, block_number, block_hash, created_at
 		FROM proof_sets
 		WHERE set_id = $1
 	`
@@ -71,15 +63,15 @@ func (p *PostgresDB) FindProofSet(ctx context.Context, setId int64, includeHisto
 	}
 	defer rows.Close()
 
-	var proofSets []*processor.ProofSet
+	var proofSets []*models.ProofSet
 	for rows.Next() {
-		ps := &processor.ProofSet{}
+		ps := &models.ProofSet{}
 		var totalFeePaidInt int64
 		err := rows.Scan(
 			&ps.ID, &ps.SetId, &ps.Owner, &ps.ListenerAddr, &ps.TotalFaultedPeriods,
-			&ps.TotalDataSize, &ps.TotalRoots, &totalFeePaidInt,
+			&ps.TotalProvedRoots, &ps.TotalDataSize, &ps.TotalRoots, &totalFeePaidInt,
 			&ps.LastProvenEpoch, &ps.NextChallengeEpoch, &ps.TotalTransactions,
-			&ps.IsActive, &ps.BlockNumber, &ps.BlockHash, &ps.PreviousID,
+			&ps.IsActive, &ps.BlockNumber, &ps.BlockHash, &ps.CreatedAt,
 		)
 		if err != nil {
 			return nil, fmt.Errorf("failed to scan proof set: %w", err)
@@ -89,37 +81,6 @@ func (p *PostgresDB) FindProofSet(ctx context.Context, setId int64, includeHisto
 	}
 
 	return proofSets, nil
-}
-
-// UpdateProofSet updates an existing proof set while maintaining version history
-func (p *PostgresDB) UpdateProofSet(ctx context.Context, proofSet *processor.ProofSet) error {
-	tx, err := p.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	// Get the ID of the current version to maintain history
-	var previousID = proofSet.ID
-
-	// Insert the new version
-	_, err = tx.Exec(ctx, `
-		INSERT INTO proof_sets (
-			set_id, owner, listener_addr, total_faulted_periods, total_data_size,
-			total_roots, total_fee_paid, last_proven_epoch, next_challenge_epoch,
-			total_transactions, is_active, block_number, block_hash, previous_id
-		) VALUES (
-			$1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14
-		)
-	`, proofSet.SetId, proofSet.Owner, proofSet.ListenerAddr, proofSet.TotalFaultedPeriods,
-		proofSet.TotalDataSize, proofSet.TotalRoots, proofSet.TotalFeePaid,
-		proofSet.LastProvenEpoch, proofSet.NextChallengeEpoch, proofSet.TotalTransactions,
-		proofSet.IsActive, proofSet.BlockNumber, proofSet.BlockHash, previousID)
-	if err != nil {
-		return fmt.Errorf("failed to insert updated proof set: %w", err)
-	}
-
-	return tx.Commit(ctx)
 }
 
 // DeleteReorgedProofSets removes proof sets from reorged blocks
@@ -150,317 +111,17 @@ func (p *PostgresDB) CleanupFinalizedProofSets(ctx context.Context, currentBlock
 				AND EXISTS (
 					SELECT 1
 					FROM proof_sets newer
-					WHERE newer.previous_id = ps.id
-						AND newer.id IN (SELECT id FROM latest_versions)
+					WHERE newer.id IN (SELECT id FROM latest_versions)
 						AND is_block_finalized(newer.block_number, $1)
+						AND newer.set_id = ps.set_id
 				)
 		)
-		DELETE FROM proof_sets ps
-		USING finalized_duplicates fd
-		WHERE ps.id = fd.id`,
+		DELETE FROM proof_sets
+		WHERE id IN (SELECT id FROM finalized_duplicates)`,
 		currentBlockNumber)
 
 	if err != nil {
 		return fmt.Errorf("failed to cleanup finalized proof sets: %w", err)
 	}
 	return nil
-}
-
-// IncrementTotalRoots increments the total_roots and total_data_size for a proof set
-func (db *PostgresDB) IncrementTotalRoots(ctx context.Context, setId int64, amount int64, totalDataSize int64, timestamp time.Time) error {
-	proofSet, err := db.FindProofSet(ctx, setId, false)
-	if err != nil {
-		return fmt.Errorf("failed to find proof set: %w", err)
-	}
-
-	proofSet[0].TotalRoots += amount
-	proofSet[0].TotalDataSize += totalDataSize
-	proofSet[0].UpdatedAt = timestamp
-
-	if err := db.UpdateProofSet(ctx, proofSet[0]); err != nil {
-		return fmt.Errorf("failed to update proof set: %w", err)
-	}
-
-	return nil
-}
-
-// DecrementTotalRoots decrements the total_roots and total_data_size for a proof set
-func (db *PostgresDB) DecrementTotalRoots(ctx context.Context, setId int64, amount int64, totalDataSize int64, timestamp time.Time) error {
-	proofSet, err := db.FindProofSet(ctx, setId, false)
-	if err != nil {
-		return fmt.Errorf("failed to find proof set: %w", err)
-	}
-
-	proofSet[0].TotalRoots -= amount
-	proofSet[0].TotalDataSize -= totalDataSize
-	proofSet[0].UpdatedAt = timestamp
-
-	if err := db.UpdateProofSet(ctx, proofSet[0]); err != nil {
-		return fmt.Errorf("failed to update proof set: %w", err)
-	}
-
-	return nil
-}
-
-// MarkProofSetDeleted marks a proof set as deleted by updating its fields
-func (db *PostgresDB) MarkProofSetDeleted(ctx context.Context, setId int64, blockNumber uint64, blockHash string, timestamp time.Time) error {
-	query := `
-		WITH old_version AS (
-			SELECT *
-			FROM proof_sets
-			WHERE set_id = $1
-			ORDER BY block_number DESC
-			LIMIT 1
-		)
-		INSERT INTO proof_sets (
-			set_id,
-			owner,
-			listener_addr,
-			total_roots,
-			total_data_size,
-			is_active,
-			next_challenge_epoch,
-			last_proven_epoch,
-			created_at,
-			updated_at,
-			block_number,
-			block_hash,
-			previous_id
-		)
-		SELECT
-			set_id,
-			'0x0000000000000000000000000000000000000000', -- zero address
-			listener_addr,
-			0, -- total_roots
-			0, -- total_data_size
-			false, -- is_active
-			0, -- next_challenge_epoch
-			0, -- last_proven_epoch
-			created_at,
-			$4, -- updated_at
-			$2, -- block_number
-			$3,  -- block_hash
-			id
-		FROM old_version;`
-
-	commandTag, err := db.pool.Exec(ctx, query, setId, blockNumber, blockHash, timestamp)
-	if err != nil {
-		return fmt.Errorf("failed to mark proof set as deleted: %w", err)
-	}
-
-	if commandTag.RowsAffected() == 0 {
-		return fmt.Errorf("no proof set found")
-	}
-
-	return nil
-}
-
-// MarkProofSetEmpty marks a proof set as empty by updating its fields
-func (db *PostgresDB) MarkProofSetEmpty(ctx context.Context, setId int64, blockNumber uint64, blockHash string, timestamp time.Time) error {
-	query := `
-		WITH old_version AS (
-			SELECT *
-			FROM proof_sets
-			WHERE set_id = $1
-			ORDER BY block_number DESC
-			LIMIT 1
-		)
-		INSERT INTO proof_sets (
-			set_id,
-			owner,
-			listener_addr,
-			total_roots,
-			total_data_size,
-			is_active,
-			next_challenge_epoch,
-			last_proven_epoch,
-			created_at,
-			updated_at,
-			block_number,
-			block_hash,
-			previous_id
-		)
-		SELECT
-			set_id,
-			owner,
-			listener_addr,
-			total_roots,
-			total_data_size,
-			true,
-			0,
-			0,
-			created_at,
-			$4,
-			$2,
-			$3,
-			id
-		FROM old_version;`
-
-	commandTag, err := db.pool.Exec(ctx, query, setId, blockNumber, blockHash, timestamp)
-	if err != nil {
-		return fmt.Errorf("failed to mark proof set as empty: %w", err)
-	}
-
-	if commandTag.RowsAffected() == 0 {
-		return fmt.Errorf("no proof set found")
-	}
-
-	return nil
-}
-
-// UpdateNextChallengeEpoch updates the next_challenge_epoch for a proof set
-func (db *PostgresDB) UpdateNextChallengeEpoch(ctx context.Context, setId int64, epoch int64, blockNumber uint64, blockHash string) error {
-	query := `
-		WITH old_version AS (
-			SELECT *
-			FROM proof_sets
-			WHERE set_id = $1
-			ORDER BY block_number DESC
-			LIMIT 1
-		)
-		INSERT INTO proof_sets (
-			set_id,
-			owner,
-			listener_addr,
-			total_roots,
-			total_data_size,
-			is_active,
-			next_challenge_epoch,
-			last_proven_epoch,
-			created_at,
-			updated_at,
-			block_number,
-			block_hash,
-			previous_id
-		)
-		SELECT
-			set_id,
-			owner,
-			listener_addr,
-			total_roots,
-			total_data_size,
-			is_active,
-			$2, -- next_challenge_epoch
-			last_proven_epoch,
-			created_at,
-			NOW(),
-			$3, -- block_number
-			$4, -- block_hash
-			id
-		FROM old_version;`
-
-	commandTag, err := db.pool.Exec(ctx, query, setId, epoch, blockNumber, blockHash)
-	if err != nil {
-		return fmt.Errorf("failed to update next challenge epoch: %w", err)
-	}
-
-	if commandTag.RowsAffected() == 0 {
-		return fmt.Errorf("no proof set found")
-	}
-
-	return nil
-}
-
-// UpdateProofSetOwner updates the owner of a proof set and maintains version history
-func (db *PostgresDB) UpdateProofSetOwner(ctx context.Context, setId int64, newOwner string, blockNumber uint64, blockHash string) error {
-	query := `
-		WITH old_version AS (
-			SELECT *
-			FROM proof_sets
-			WHERE set_id = $1
-			ORDER BY block_number DESC
-			LIMIT 1
-		)
-		INSERT INTO proof_sets (
-			set_id,
-			owner,
-			total_roots,
-			total_data_size,
-			total_faulted_periods,
-			total_proved_roots,
-			total_fee_paid,
-			total_transactions,
-			is_active,
-			next_challenge_epoch,
-			last_proven_epoch,
-			created_at,
-			updated_at,
-			block_number,
-			block_hash,
-			listener_addr,
-			previous_id
-		)
-		SELECT
-			set_id,
-			$2, -- new owner
-			total_roots,
-			total_data_size,
-			total_faulted_periods,
-			total_proved_roots,
-			total_fee_paid,
-			total_transactions,
-			is_active,
-			next_challenge_epoch,
-			last_proven_epoch,
-			created_at,
-			NOW(),
-			$3, -- block_number
-			$4, -- block_hash
-			listener_addr,
-			id
-		FROM old_version;`
-
-	commandTag, err := db.pool.Exec(ctx, query, setId, newOwner, blockNumber, blockHash)
-	if err != nil {
-		return fmt.Errorf("failed to update proof set owner: %w", err)
-	}
-
-	if commandTag.RowsAffected() == 0 {
-		return fmt.Errorf("no proof set found")
-	}
-
-	return nil
-}
-
-// UpdateProofSetStats updates the proof stats for a proof set
-func (db *PostgresDB) UpdateProofSetStats(ctx context.Context, setId int64, proofsSubmitted int64, periodsFaulted int64, blockNumber uint64, blockHash string, timestamp time.Time) error {
-	tx, err := db.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	proofSets, err := db.FindProofSet(ctx, setId, false)
-	if err != nil {
-		return fmt.Errorf("failed to find proof set: %w", err)
-	}
-
-	if len(proofSets) == 0 {
-		return fmt.Errorf("no proof set found")
-	}
-
-	proofSet := proofSets[0]
-	previousID := proofSet.ID
-
-	// Insert new version with updated stats
-	_, err = tx.Exec(ctx, `
-		INSERT INTO proof_sets (
-			set_id, owner, listener_addr, total_faulted_periods,
-			total_data_size, total_roots, total_proved_roots, total_fee_paid,
-			last_proven_epoch, next_challenge_epoch, total_transactions,
-			is_active, block_number, block_hash, previous_id
-		) VALUES (
-			$1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15
-		)
-	`, proofSet.SetId, proofSet.Owner, proofSet.ListenerAddr,
-		proofSet.TotalFaultedPeriods+periodsFaulted, proofSet.TotalDataSize,
-		proofSet.TotalRoots, int64(proofSet.TotalProvedRoots)+int64(proofsSubmitted), proofSet.TotalFeePaid,
-		blockNumber, proofSet.NextChallengeEpoch,
-		proofSet.TotalTransactions, proofSet.IsActive,
-		blockNumber, blockHash, previousID)
-	if err != nil {
-		return fmt.Errorf("failed to insert proof set: %w", err)
-	}
-
-	return tx.Commit(ctx)
 }
