+++ backend/indexer/internal/models/provider.go
@@ -1,0 +1,15 @@
+package models
+
+import (
+	"time"
+)
+
+type Provider struct {
+	ReorgModel
+	Address             string    `db:"address" json:"address"`
+	TotalFaultedPeriods int64    `db:"total_faulted_periods" json:"total_faulted_periods"`
+	TotalDataSize       int64    `db:"total_data_size" json:"total_data_size"`
+	ProofSetIds         []int64   `db:"proof_set_ids" json:"proof_set_ids"`
+	CreatedAt           time.Time `db:"created_at" json:"created_at"`
+	UpdatedAt           time.Time `db:"updated_at" json:"updated_at"`
+}
